libutil/fileref: new class
Problem: an object similar to the KVS treeobj objects is
needed for representing references to files whose data may
be mapped into the content store.

Introduce a simple class for 'fileref' objects.  The object
consists of file metadata plus an array of blobrefs for data.
It can represent

- regular files
- directories
- symbolic links

Sparse files are represented compactly, which may be useful for
mapping file system images for virtual machines.

If a file contains less than 4K of data, the data is represented
within the fileref rather than requiring indirection through the
content cache.