shell: split load of initrc from shell.init callback
Problem: The shell initrc (and thus any dynamically loaded plugins)
is loaded in the same function as the shell.init plugin callback, but
this makes it impossibke to call other functions between these two
operations.

Split the current shell_init() function into a shell_initrc(), where
the initrc is loaded, and a shell_init(), where only the shell.init
plugin callbacks are invoked for future flexibility.