shell: use taskmap instead of rcalc basis for task mapping
Problem: The job shell is restricted to block distribution mapping of
taskids to nodes because of the use of a single "basis" id for each
shell rank in the shell's rcalc code.

Remove the shell's use of the rcalc rankinfo 'basis' completely, and
replace with a block distribution taskmap object and corresponding
taskid idset generated from the taskmap.

Use the taskids idset wherever the list of local taskids is required,
e.g. when initially creating task, when logging the task id set for
debugging, and when generating the rank information JSON object for
flux_shell_get_rankinfo(3).