shell: move shell.init event context init after initrc
Problem: The call to register the initial shell.init event context
comes before the shell initrc is loaded, which, when modificiation
of the shell taskmap is enabled, may change the contents of that
context.

Move shell_register_event_context() to _after_ shell_initrc() so
that any changes to the taskmap will be reflected in the shell.init
event context.