broker/content-cache: revalidate mapped blobs
Problem: if a file is modified while it is mapped into the content
cache, hashes could become invalid without detection.

When a mapped blob is accessed from the rank 0 cache, check that the
hash still matches the content and throw an error if not.  This check
is only performed for mapped cache entries, and only on rank 0.