job-manager: add queue-start rpc target
Problem: The logic for enabling/disabling queues is maintained
in `job-manager/queue` and logic for scheduling queues is
maintained in `job-manager/alloc`.  It would be nicer if they
were in the same place.

Add a new rpc target `job-manager.queue-start` that will be able
to manage start/stop of queues.