shell: support taskmap shell option
Problem: The job shell does not support alternate task mapping
mechanisms.

Support a 'taskmap' job shell option which should be set to an object
with keys

  scheme: the alternate task mapping scheme to use. 'block' is the
  default and therefore does nothing, and 'manual' is a builtin
  scheme which sets the taskmap explicitly to 'value'. O/w, a
  'taskmap.<scheme>' plugin callback will be issued which should
  return a new taskmap.

  value: an optional value argument to the taskmap 'scheme'.

If this jobspec shell option is present, then the job shell will
attempt to call a plugin callback `taskmap.<scheme>`, passing in the
`scheme`, `value`, and encoded current taskmap (in block distribution)
to the callback.

The plugin callback should then push a new RFC 34 encoded task mapping
onto the plugin out arguments. If this mapping is valid, then the
shell will replace the existing mapping with the plugin-provided one.

The shell also supports a `manual` scheme, in which the taskmap can
be provided directly in the jobspec as taskmap.value. Since the shell
maps tasks in block distirbution by default, a scheme of "block" is
a noop.