libutil/hola: add hash-of-list class
Problem: implementing a hash-of-lists using off the shelf hash and list
functions is straightforward, but results in boiler plate code that could
benefit from being reusable and separately testable.

Add a new class that provides this abstraction using zhashx_t and zlistx_t
internally.  The interfaces are identical to the APIs of those classes
where applicable to reduce the learning curve and leverage existing
zhashx(3) and zlistx(3) documentation.