shell: replace task-count with taskmap in shell.start event
Problem: The mapping of taskids to nodes is not emitted in the job
record by the job shell, and thus determination of which node a given
taskid was run on is fragile at best.

Record the RFC 34 taskmap object in the shell.start event of the exec
eventlog. This allows users or tools to recreate the precise mapping
of task ranks to nodes any time after the job shell has started.

Remove the now redundant 'task-count' member of the shell.start
event context.