flux-mini: rename and generalize Dependency class
Problem: The class used in flux-mini to parse dependency option
arguments as URIs is actually more general than the Dependency class
name would imply.

Rename the Dependency class to URIArg so that it can be used for any
uri-style option argument. Add a 'name' argument to the class so that
error messages can be speific to the option being procesed. Extend
the class to allow the `:` character to be optional, in cases where
a scheme alone is allowed to be specified.