job-manager: support per queue start/stop
Problem: Individual queues cannot be started/stopped.

Solution: Update the queue-start RPC to allow individual
queues to be started/stopped.  When a queue is stopped, alloc
requests for jobs in that queue are canceled, and
the jobs are removed from the alloc queue and prevented
from re-entering it.

Adjust test in t2240-queue-cmd.t for change in test output.

Fixes #4629