broker: fix error when broker.mapping attr is not set
Problem: If the broker.mapping attr is not set, as in when the
PMI_process_mapping key overflows the max PMI value size, then
the derivation of broker.critical-ranks fails which causes a fatal
error. But Flux should be able to bootstrap without broker.mapping.

Do not issue a fatal error when the broker critical ranks can't be
calculated. Instead issue a warning and continue. The end result will
be that the parent will treat all broker ranks as critical.

Fixes #4777