broker: encode broker.mapping attribute as RFC 34 taskmap
Problem: The broker.mapping attribute is stored in the PMI-1
PMI_process_mapping format documented in RFC 13, but using a Flux
Task Map as documented in RFC 34 would be more convenient since a
library is provided for working with this encoding.

Encode broker.mapping as an RFC 34 Flux Task Map. Update users
of this mapping to decode using libtaskmap.