flux-start: add --recovery option
Problem: sometimes a Flux system instance will refuse to start, and
then debugging is tricky because the system cannot be interactively
probed.

Add flux start --recovery[=ARG], which starts a singleton instance
using state from a previous instance.

If ARG is unspecifed, recover the system instance, e.g.
  sudo -u flux flux start --recovery

If ARG is a directory, recover a persistent 'statedir', e.g.
  flux start --recovery=/tmp/statedir

If ARG is a file, recover a flux-dump(1) archive, e.g.
  flux start --recovery=/tmp/mydump.tar
or for a system dump:
  sudo -u flux flux start \
    --sysconfig --recovery=/var/lib/flux/dump/20221127_065818.tgz

Fixes #4466