job-manager: add queue to job data structure
Problem: In the future we will need to know if a job is in a queue, but it
is inconveniently not a part of the 'job' data structure in job-manager.

Solution: Add a 'queue' field to the 'job' data structure.  Parse it
in job_create_from_eventlog() and job_create_from_json().  Add unit
test.