broker: ignore rc1 failure in recovery mode
Problem: recovery mode aborts if something goes wrong in rc1,
such as a module failing to load.

This is the primary reason why a system instance wouldn't start,
so if rc1 fails in recovery mode, post a new "rc1-ignorefail" event
that transitions to the quorum state instead of shutdown.