resource: use R node count to make up/down sets
Problem: in recovery mode, if R contains more execution targets
than the instance size, as may be the case if the instance was
launched with PMI, the extra targets are marked "up" and the
scheduler thinks it can allocate them to pending jobs.

In the resource module, use the execution target count in R, if
available, instead of the instance size in calculating the set of "up"
and "down" rank sets.