job-manager: support nocheckpoint for queue start
Problem: When a flux instance is shutdown, its common to stop
scheduling during the shutdown process.  Since scheduling is
checkpointed and saved for restart, this means that a queue could
always be stopped when an instance is restarted.

Solution: Add a "nocheckpoint" flag to the job-manager's queue-start
rpc.  If it is set to true, do not checkpoint that change to the
queue state.  Use a "shadow value" to hold the value that should be
checkpointed, which may be different than the real start state of
a queue.