etc/rc1: use --nocheckpoint when stopping queue
Problem: When a flux instance is shut down, it is common to stop
the scheduling of jobs and waits for current jobs to be canceled.
The stopping of the queue can be saved as a checkpoint and will be
assumed to be stopped if the flux instance is restarted.  This may
not be desireable because stopping the queue during shutdown is
not a "normal" stopping of the queue.

Solution: Add --nocheckpoint to etc/rc1 and t/rc/rc1-job to avoid
checkpointing during a "flux queue stop".