job-manager: perilog: destroy kill future on timeout
Problem: The job-manager perilog plugin does not destroy the subprocess
kill future after the future times out. This could cause the future
callback to be called continuously, which is not only unnecessary,
but could be dangerous and may be a proximate cause of the broker
crash in #4790, if somehow the subprocess handle becomes invalid
before one of the callbacks is made.

Destroy the kill future on ETIMEDOUT in the perilog plugin so that the
kill future callback does not get unnecessarily called more than once.

Fixes #4790