testsuite: ensure a job with a prolog can timeout
Problem: In issue #4790 a broker crash occurred in the perilog plugin
after a job receives a timeout exception, but this case isn't tested
in the testsuite.

While there should not be a difference between a cancel (already
tested) and timeout exception (untested), it seems worthwhile to
explicitly test the timeout case. Therefore, add a test which ensures
a job can get a timeout exception without a crash when job manager
prologs are active.