github: use ubuntu-20.04 for jobs that break on ubuntu-latest
Problem: The ubuntu-latest image seems to have been recently updated
to Ubuntu 22.04, which broke the python linting job, which now does't
support python 3.6. Additionally, the system/coverage test build fails
because docker-run-systest.sh apparently breaks on 22.04.

Use runs-on: ubuntu-20.04 in the python-lint and ci-checks jobs to work
around the failures for now.