broker: fix `flux start bash`
Problem: The broker currently can only determine that rc2 is
interactive when no command is provided, however a shell with no
other arguments should also be considered interactive. Since this
is not the case, `flux start $SHELL` hangs. This occurs because the
broker runs all initial programs it considers non-interactive in a
new process group.

Detect if rc2 is interactive by comparing any single argument to
entries in /etc/shells using getusershell(3).