broker: allow shells other than $SHELL
Problem: There is no way to pass a shell other than $SHELL as the
broker rc2 command, since runat_command_set_cmdline() does not take a
shell argument. This makes it so that `flux start csh` does not work
when SHELL=/bin/bash, for example.

Add an optional shell argument to runat_push_shell() and
runat_command_set_cmdline(). When running an interactive shell in
the broker, push the actual argv in case the user provided a shell
other than the default.