broker: kill non-interactive subprocesses that are stopped
Problem: If the broker runs a process that appears to be
non-interactive which then tries to access the tty, it will be
stopped due to SIGTTIN since the broker runs these processes in a
new process group. This appears as an uninterruptible hang.

Detect when a broker child is stopped and assume this is an error
condition. Issue an error then kill the child to avoid a hang.