testsuite: ensure that broker kills stopped child
Problem: There is no test that ensure the broker behaves as expected
when an interactive subprocess run in non-interactive mode is stopped
by a signal, presumably due to accessing the tty.

Add issues/t4771-flux-start-bash.sh which mimics the execution of a
non-interactive rc2 process accessing the terminal, by running bash
as a symlink. In the test, ensure the rc2 process is killed instead of
hanging.