shell: pmi: set flux.taskmap PMI key to RFC 34 taskmap
Problem: When mapping tasks with cyclic task distribution and a high
number of tasks per node, PMI_process_mapping will overflow the max
PMI value length of 1024. This causes Flux's libpmi to be unable to
calculate clique information.

Add a flux.taskmap key to the job shell PMI kvs which encodes the task
mapping in RFC 34 Task Map form, which can encode cyclic task mappings
more compactly. When Flux is a PMI client, it will then be able to
check for this key and thereby handle a wider range of mappings.