libpmi: check for flux.taskmap key before PMI_process_mapping
Problem: A PMI_process_mapping string for a cyclic task distribution
of greater than about 126 tasks over more than one node will overflow
the maximum PMI value length, and therefore clique information for
task distributions of this type and size cannot be computed by libpmi.
However, the Flux job shell also places an RFC 34 Task Map in the
flux.taskmap key, which should be able to handle a wider variety of
encodings.

In the libpmi simple client, check for a flux.taskmap key and use that
as the job taskmap instead of PMI_process_mapping when it exists.