libhostlist: fix 8K limitation on encoded hostlist
Problem: libhoslist uses a static 8K buffer in hostlist_encode()
to hold the encoded hostlist string, but this places an unnecessary
limit on the size of a hostlist that can be encoded.

Switch the internal hostlist_encode() implementation to use a POSIX
memstream instead of a static buffer so that a hostlist of arbitrary
size can be encoded. This also simplifies the code since bounds
checking can be removed.

To simplify the resulting code and avoid copying a buffer, drop the
use of hostrange_numstr() and directly print the hostrange range
(if any) into the memstream.

Fixes #4702