shell/stage-in: change destination option
Problem: there is no way to specify that the stage-in destination
is a parallel file system.

Rename the stage-in.destdir option to stage-in.destination and
allow an optional "global:" or "local:" prefix.

If "global:" is specified, the copy is only performed on shell rank 0.
If "local:" is specified, the copy is performed on all ranks as before.
If no prefix is specified, local is assumed.

Rename the destination option used in t2617-job-shell-stage-in.t.

Fixes #4816.