libpmi: add universal PMI client library
Problem: flux-pmi needs the ability to dlopen a PMI library
so that we can add tests for the installed libpmi.so, but to do
so would require tricky code in the broker to be replicated.

Create a new interface that both the broker and flux-pmi can use.

Implement "plugins" (although they are all built-in) for each type
of PMI client, which avoids the confusing conditional logic and
preprocessor ifdefs that the broker used.

A URI is used to select the plugin.  If the URI argument is NULL,
then the implementations are searched in the same order the broker
searched them:  simple (wire proto), pmix (if compiled), libpmi (dlopen),
and single.  The optional URI path component is passed to the plugin.
For example, the libpmi plugin uses the path to override the dlopen()
filename parameter.

In the future perhaps external plugins could be supported.  For example,
upmi_pmix could be supplied by the flux-pmix project and flux-core would
not have to be conditionally compiled with openpmix.  For this step, I
just wanted to complete the refactoring in a direct way.