broker: use libupmi
Problem: the broker has its own "pmiutil" abstraction to allow it to
launch under Flux and various foreign launchers, but this abstraction
is not shared with other parts of Flux.

Convert the broker to use the internal libupmi universal pmi library
instead of its one-off one.

The behavior should be mostly the same as before, as libupmi searches
its plugins in the same order as pmiutil.  A minor difference is that
the tracing output with FLUX_PMI_DEBUG=1 has changed slightly.

The code for subverting dlopen()'s library search to skip Flux's libpmi.so
does not appear to be needed anymore and is removed. The Flux broker should
never find the libpmi.so since
- It is installed to a side directory, so it should never be found when a
  foreign launcher starts Flux, or when 'flux start' execs the broker
  standalone.
- Although the side directory is added to LD_LIBRARY_PATH in the Flux job
  environment (by MPI plugins), when Flux launches Flux, or when flux-start
  launches flux with the --test-size option, it provides the simple wire
  protocol, which is found before the library search is attempted.
Set the UPMI_LIBPMI_NOFLUX flag so that if the broker ever does find Flux's
libpmi.so, it will fail and will try the next upmi plugin ('single').