build: construct libpmi.so in a subdir
Problem: dlopen() always finds libpmi.so when Flux is running
in the source tree.

When running in-tree, LD_LIBRARY_PATH includes src/common.
Since libpmi.so is also built there, it is found by dlopen().

Construct the library a sub-directory instead (src/common/flux)
to make the in-tree behavior more like the installed behavior.