upmi_libpmi: ignore PMI_LIBRARY in environment
Problem: PMI client takes a dlopen() argument directly from the
environment, potentially allowing a user to execute arbitrary code.

This was flagged by CodeQL as CWE-114:
https://cwe.mitre.org/data/definitions/114.html

I don't think this is actually a problem since the broker is typically
launched in an environment that is either set up by the same user or
by the flux user.  However, given that this feature was not documented
with other PMI_* environment variables in RFC 13 and that it is of limited
usefulness, just drop it.