flux-top: Support --color option
Problem: By default flux-top outputs certain fields with color.
However, this can be problematic in some user's terminals.

Solution: Support a --color option that is very similar to the one
in flux-dmesg and flux-jobs.  This will allow users to disable color
if they want.  Color choices can be auto, always, or never.

Fixes #4835