libjob: add flux_job_timeleft
Problem: It is tricky for a process to determine the current remaining
time in a Flux job with which it is associated, since the process
could be part of a job or part of an initial program within a job. Flux
should provide a function to hide this complexity.

Add flux_job_timeleft(3) which, given a Flux handle to the current
instance, determines if the current process in a job or initial
program context, fetches the correct expiration and calculates the
correct remaining time.