python: add assumeFuture argument to flux.util.parse_datetime
Problem: The parse_datetime() utility function is statically configured
to assume future dates when a term like "Friday" is given. That is,
instead of giving the date for the previous Friday, the function will
return then next Friday instead. This behavior should be configurable.

Add an assumeFuture parameter to the function which defaults to True.
If set to False, then parse_datetime() will assume dates in the past
instead.