job-manager: reduce amount of checkpointing
Problem: Everytime a queue is enabled/disabled or started/stopped,
the current queue state is commited to the KVS for checkpointing.  This
is an unnecessary amount of checkpointing, as it is
only really needed when the job-manager module is unloaded.

Solution: Do not sync queue state to the KVS everytime the queue state
is modified.  Only do this when the module is unloaded.

Fixes #4804