testsuite: update tests for default queue settings
Problem: Many tests assume newly configured queues default
to being enabled/started.  However this is no longer true.

Solution: Update tests to adjust for the fact queues are
enabled/stopped by default.  In most cases, a call to
 "flux queue start" after a new queue is configured or update
expected output.  In some cases, we must adjust tests because
a queue begins in the stopped state instead of the started state.