testsuite: cover more queue states on restart
Problem: In t2219-job-manager-restart, only queue states of "disable"
and "stop" are covered when testing that queue state is preserved
on restart.  It would be nice to also cover "enable" and "start".

Solution: Enhance tests to cover both the "enable" and "start" queue
state as well as "disable" and "stop".