libupmi: convert to flux_plugin_t
Problem: current libupmi plugin structure cannot handle plugins
provided by external projects, and pmix support is to be moved
to the flux-pmix project.

Rework the libupmi internals to use flux_plugin_t instead of a struct
of callbacks with custom C signatures.

Add a new callback called "preinit" which is invoked after the plugin
is created to initialize it and check whether it is a good choice for
a bootstrap mechanism, without committing to using it.

The common methods are built in, but a URI scheme may now be resolved
by a dso found in the upmi module search path, which is determined by
- compiled in ${libdir}/flux/upmi (or a source tree alternative), or
- the FLUX_PMI_CLIENT_SEARCHPATH environment variable, if set

In this application, the order that methods are tested is significant.
The order is determined by
- compiled in list "simple libpmi single", or
- the FLUX_PMI_CLIENT_METHODS environment variable, if set