flux-resource: support output field "propertiesx"
Problem: It is very common to assign the name of a queue as
a "property" for a set of resources.  When both queues and
properties are listed under "flux resource list", it would be
redundant to see queue name listed under both "queue" and
"properties".

Solution: Support a new output field called "propertiesx", which
is the same as "properties" but will not list any properties that
are also queues.