testsuite: make test independent on default format
Problem: In t2315-resource-system, there is a test that compares
against the default output format of "flux resource list".  As this
format can change in time, it is best for the test to be independent
of that format.

Solution: Use the --format option to make the test independent of the
default format.