job-manager: switch to timer watcher in perilog plugin
Problem: A rare segfault seems to be triggered by the job manager
'perilog' plugin. Coredump analysis shows that when an internal
subprocess kill future is fulfilled with ETIMEDOUT, the perilog_proc
structure is invalid, resulting in a segfault when the plugin tries to
send SIGKILL to the prolog subprocess. It is unknown why this occurs,
since the future should be destroyed when the `struct perilog_proc`
is, therefore the future should never be fulfilled if the perilog_proc
entry has been freed.

However, it may make the code more clear and robust if the kill timeout
is implemented with a traditional timer watcher, instead of abusing
the timeout of the original kill future to trigger SIGKILL. Perhaps
something in this use case is causing a race when the process is
exiting or has exited. Therefore, switch the prolog kill timeout to
use a traditional timer watcher in hopes this will eliminate any race
or other unknown issue in the code.

If we still see a segfault in this area of code, perhaps it will
be easier to diagnose with the timeout handled separately from the
kill future.

Fixes #4790