cmd: add flux-pgrep and flux-pkill
Problem: There is not a convenient way for users to list and/or cancel
jobs based on a pattern match of the job name. Additionally, users
have requested a way to query or cancel jobs with job ids in a range.

Add a simple flux-pgrep(1) command inspired by the well-known pgrep(1)
utility. By default this command prints the job ids of all jobs whose
name matches a supplied pattern. The command additionally supports a
`jobid1..jobid2` syntax to indicate that a range of jobids are to be
matched. If a pattern and id range are both supplied, then they both
have to match.

If flux-pgrep is called with the name flux-pkill, then matching jobs
are canceled instead of listed.