docker: add site-packages to default python3 path
Problem: On debian based docker containers (bionic, focal), the flux
python module is not found by default because

  /usr/lib/python3.x/site-packages

is not in the default search path. This is because debian based
distributions only search 'dist-packages' instead of 'site-packages'.
However, `make install` puts our Flux modules in 'site-packages'
because this is the module install directory understood by autotools.

To work around this issue, place a 'site-packages.pth' file in
'dist-packages' to get Python to read site-packages as well as
dist-packages.

Fixes #4877