flux-keygen: add --meta KEY=VAL,... option
Problem: it may be useful to allow certificate meta values to
be overridden, or for for arbitrary new metadata to be added
to the certificate.

Add a new --meta KEY=VAL option.  The option may be specified multiple
times, or multiple KEY=VAL pairs may be specified with the same option,
separated by commas.

Fixes #4875