libflux: fix message encoding problem
Problem: messages are encoded with some integers in little endian
byte order, but RFC 3 requires network byte order (big endian).

This bug was introduced by the cleanup in #4823 where the libccan
pushpull class was used for encoding integers in the PROTO frame.
Unfortunately, it was not noticed that these functions encode to
the wrong byte order.

This bug appeared in the v0.46.1 release and will be removed in v0.47.

Fix the proto_encode() and proto_decode() to use the correct order.

Fixes #4886