testsuite: check for changes in message encoding
Problem: we don't currently have a way to ensure that message
encoding doesn't change from version to version.

Add a tool that encodes a stream of sample messages on stdout.
Add the current known good encoded output to marshall/valid/v0.47.data.
For what it's worth, add the known bad encoded output of v0.46.1
to marshall/invalid/v0.46.1.data.  Add a sharness test that ensures
that the encoding doesn't change.