flux-resource: do not assume "requires" config
Problem: flux-resource currently assumes that every queue has
a "requires" configuration set.  However, it is not required, which
can lead to errors.

Do not assume "requires" is set for every queue, if it is not set,
then the queue does not have any requirements and applies to all
resources.

Fixes #4891