flux-resource: do not set queue if resources empty
Problem: It can be strange to see a queue on a resources line if
no resources exist in that state.  For example:

     STATE QUEUE    NNODES   NCORES    NGPUS NODELIST
      free all,batc      2        2        0 node[1-2]
      free all,debu      2        2        0 node[3-4]
 allocated all           0        0        0
      down all           0        0        0

The queue "all" is listed for the allocated and down nodes, which there are
no resources in that state.

Solution: If there are no resources assigned in the current line, do not set
a queue.