flux-mini: update --requires to use constraint.parser
Problem: The flux-mini --requires option does not support the constraint
syntax documented in RFC 35.

Add a subclass of flux.constraint.parser.ConstraintParser customized for
job constraints and use this parser in handling the --requires option.
This means that the --requires option now supports the full RFC 35 syntax
plus the following:

 - the default operator is "properties" so that previous behavior of
   --requires is preserved, e.g. --requires=foo requires property "foo".
 - The properties operator has its values split on comma by default to
   preserve previous behavior of --requires, e.g. --requires=foo,bar
   results in {"properties": ["foo","bar"]}.
 - host: hosts: and hostlist: may all be used to require or exclude
   resources by hostname
 - rank: or ranks: may be used to require or exclude resources by
   broker rank.