libschedutil: handle hello failure gracefully
Problem: if scheduler cannot reallocate resources to a running job,
the scheduler interface is torn down, requiring sys admin intervention.

This was seen in conjunction with flux-framework/flux-sched#992.

It's not really necessary for this to be fatal to the instance.
Raise a fatal exception on the job and let it be cleaned up in the
usual way.