job-manager: handle scheduler-restart exception
Problem: after a scheduler-restart exception, the job manager
still thinks the job is holding resources that it must free.

Clear the has_resources flag on the job when this exception is
posted so that it won't get stuck in CLEANUP and/or confuse the
scheduler with a free request.  The resources are effectively
revoked.