frobnicator: support full range of input job constraints
Problem: The job frobnicator constraints plugin assumes that user job
constraints will only contain properties, but constraints may now contain
multiple operators including nested logical expressions.

Update the frobnicator constraints plugin to join the user and queue
constraints with the "and" operator if the top-level constraint operator
is not "properties". If "properties" is the top-level operator then
continue to append the required queue properties to the values array.

Update one frobnicator test that assumes the order of constraint properties.