python: support 'i' special presentation type
Problem: In the future, some listing output may output "inf" (infinity)
because that is a common default if no configuration is done.  Outputting
"inf" is fine, but outputting something else would be convenient to let
the user know nothing was configured.

Solution: Support a new special presentation type "i", which will turn
"inf" to a hyphen.