alloc-bypass: emit set-flags event in job.validate
Problem: The alloc-bypass jobtap plugin is currently not functional
because the set-flags event is emitted in job.state.sched. Since the
job manager no longer processes events as they are emitted, events
posted during a state callback are not handled until after that state
has been processed. Thus the job manager issues an alloc request to the
scheduler before the set-flags event to add alloc-bypass is processed.

Move the call to emit the set-flags event to the job.validate callback.
This will be early enough such that the flag will be processed before
the the SCHED state and alloc request.

Fixes #4900