testsuite: fix alloc-bypass tests
Problem: t2273-job-alloc-bypass.t doesn't properly test that an
alloc-bypass job does not result in an alloc request to the scheduler.
This is because a full-system job is run first, which means the
scheduler has no resources to allocate to the bypass job.

Add a test case which runs an alloc-bypass job *first* on an empty
system, then tries to run a full system job. This should catch the
case described in issue #4900 where the scheduler allocates a core
to the alloc-bypass job.