python: default write_tables to False in ConstraintParser
Problem: The ply `parsetab.py` parser table is supposed to be
generated in the directory where the parser grammar is defined if
the table does not already exist, or if a checksum of the grammar
does not match an existing table. For the ConstraintParser, this
table should be generated during the build and becomes part of the
installation on make install.  Since the grammar will not change,
the parsetab.py should be good for the life of the installation.

However, when ConstraintParser is used via flux-mini.py via `flux
mini --requires`, the parsetab.py file is being regenerated in
${libexec}/flux/cmd, instead of the expected location.  This may be
because flux-mini subclasses ConstraintParser, but the real reason
is not known. This behavior will result in errors on every invocation
of flux-mini --requires on an installed system.

Default the `write_tables` parameter to yacc.yacc() to False to avoid
attempting to write out parsetab.py unless explicitly set to True.
Pass write_tables=True in the flux.constraint.parser __main__ function
since it is this helper that generates the file during make.