job-manger: add 'history' jobtap plugin
Problem: it's not possible to query job-list for a list of
recently submitted jobs in any state.

Running jobs are always returned first, then pending, then inactive.
Given that the number of jobs returned must be capped to avoid a huge
result, it's not possible to construct a query to job-list for the most
recently submitted jobs, even if the results are re-sorted upon receipt.

Add a simple jobtap plugin that tracks jobs for each user in submission
order.  It registers a service and answers based on python array slice
syntax for the requesting user.

Jobs are never removed from this list.
A future commit will deal with purging.