flux-job: add last subcommand
Problem: there is no simple command to obtain the job IDs of
the most recently submitted jobs.

Add "flux job last" which acts as a client for the service registered
by the 'history' jobtap plugin.  It takes one "slice" optional argument.
If not provided, the default is "[:1]" which returns the most recently
submitted job.

In addition, an argument of N is an alias for "[:N]" (most recent N jobs).