flux-resource: mark drained+offline nodes with asterisk
Problem: It is difficult to tell which drained nodes are offline vs
not in the output of flux-resource drain and status.

Take a simple step at first and flag offline drained nodes with an
asterisk, i.e. `drained*`, in flux-resource `status` and `drain`
output.

Update the expected output in one test.

Fixes #4909