libutil/slice: new class
Problem: the python array slicing syntax may be familiar to
users so it's a good candidate for user interfaces, but we have
no way to process a slice in C.

Add a "slice" class that parses a slice expression, and provides
a minimal first/next style interface for obtaining sliced "array
indicies".  The initial use case is to slice a linked list by list
position, so this seemed reasonable.  Other interfaces to slice
other data structures could be added later based on demand.