flux-job: attach: add status line for blocked jobs
Problem: When there is no output from `flux job attach`, it is
impossible for a user to tell if the job is blocked waiting for
priority, dependencies, resources, or the prolog, or if the job has
been started and is hanging.  This can be frustrating.

Add support for a status line in `flux job attach`, emitted when the
'start' event has not arrived within 2s of program startup. The
status line offers a short description of what the job is waiting
on, along with a timer that is updated every second so the user can
see the command has not hung.

This feature is disabled as unnecessary if he --show-events flag
was used.

To force the status line on immediately (and for testing), a new
--show-status flag is added to `flux job attach`.