python: restrict valid characters in RFC 35 operators
Problem: The flux.constraint.parser RFC 35 parser breaks an expression
into 'operator[:operand]' terms, but then fails to check if the
provided operator is free of invalid characters. This can result in
confusion for first-time users that try expressions like `4g,host:foo`.

Check derived operators for invalid characters. For now, operators are
only allowed to contain alphanumeric characters, `_`, `-`, `+` and `@`.

This allows a much better error message to be presented to the user, e.g.

    invalid character ',' in operator '4g,host:'