job-validator: enhance invalid job constraint detection
Problem: The jobspec job validator plugin only checks for valid
conditional and property constraints. This allows invalid hostlist
and ranks constraints, along with invalid operators, to unnecessarily
reach the scheduler.

Support validation of hostlist and ranks constraints. Also raise an
error if the operator is not one of the known RFC 31 operators.