flux-mini: fix handling of rank/hostlist terms joined by AND
Problem: The flux-mini MiniConstraintParser class adds 'hostlist'
and 'ranks' to the set of combined_terms, which means ranks:1 and
ranks:2 are combined to ranks:1-2. But 'ranks:1 and ranks:2' (which
can never match, since no node can be rank1 AND rank2) is not the
same as ranks:1-2, which matches any rank in the set 1-2.

Though it is unlikely anyone will want to ever join two hostlist/ranks
operators with AND, remove these terms from the combined_terms set
in MiniConstraintParser to avoid any confusion.